[MOVE] : article-detail-tag / #23
article detail adapter viewHolder 위치 이동
